+++ tests/Integration/Compiler/Util/Compiler.fs
@@ -32,7 +32,6 @@ module Compiler =
     let sourceFile = IO.Path.Join(projDir, "Program.fs" ) |> Path.normalizeFullPath
 
     let cliArgs =
-        Log.makeSilent()
         let compilerOptions = CompilerOptionsHelper.Make()
         { CliArgs.ProjectFile = projFile
           FableLibraryPath = None
@@ -51,7 +50,8 @@ module Compiler =
           Exclude = ["Fable.Core"]
           Replace = Map.empty
           RunProcess = None
-          CompilerOptions = compilerOptions }
+          CompilerOptions = compilerOptions
+          Verbosity = Verbosity.Normal }
 
     let mutable private state = State.Create(cliArgs, recompileAllFiles=true)